+++ src/view/day-list.js
@@ -1,0 +1,3 @@
+export const createDayListTemplate = () => {
+  return (`<ul class="trip-days"></ul>`);
+};
